Refactor event list and calendar to use the same page wrapper
Including shared navigation tabs!
