+++ app/routes/events/components/Calendar.tsx
@@ -1,16 +1,15 @@
-import { Icon, LinkButton, Page } from '@webkom/lego-bricks';
+import { Icon } from '@webkom/lego-bricks';
 import { usePreparedEffect } from '@webkom/react-prepare';
 import moment, { type Moment } from 'moment-timezone';
 import { Helmet } from 'react-helmet-async';
 import { useParams } from 'react-router-dom';
 import { fetchEvents } from 'app/actions/EventActions';
 import { useCurrentUser } from 'app/reducers/auth';
-import { useAppDispatch, useAppSelector } from 'app/store/hooks';
+import { useAppDispatch } from 'app/store/hooks';
 import createMonthlyCalendar from 'app/utils/createMonthlyCalendar';
 import styles from './Calendar.module.css';
 import CalendarCell from './CalendarCell';
 import EventFooter from './EventFooter';
-import EventsTabs from './EventsTabs';
 
 const WEEKDAYS = ['Man', 'Tir', 'Ons', 'Tor', 'Fre', 'Lør', 'Søn'];
 
@@ -35,7 +34,6 @@ const Calendar = () => {
   const { month, year } = useParams<{ month: string; year: string }>();
   const date = getDate(month, year);
 
-  const actionGrant = useAppSelector((state) => state.events.actionGrant);
   const currentUser = useCurrentUser();
   const icalToken = currentUser?.icalToken;
 
@@ -63,15 +61,7 @@ const Calendar = () => {
   );
 
   return (
-    <Page
-      title="Arrangementer"
-      actionButtons={
-        actionGrant?.includes('create') && (
-          <LinkButton href="/events/create">Lag nytt</LinkButton>
-        )
-      }
-      tabs={<EventsTabs />}
-    >
+    <>
       <Helmet title="Kalender" />
 
       <h2 className={styles.header}>
@@ -100,8 +90,9 @@ const Calendar = () => {
           />
         ))}
       </div>
+
       {icalToken && <EventFooter icalToken={icalToken} />}
-    </Page>
+    </>
   );
 };
